refactor: migrate to new data fetch module
