+++ tests/test_strategies_module.py
@@ -20,10 +20,10 @@ tzlocal_mod.get_localzone = lambda: None
 sys.modules["tzlocal"] = tzlocal_mod
 
 # Stub internal modules pulled in by bot_engine imports we don't exercise
-df_stub = types.ModuleType("ai_trading.data_fetcher")
+df_stub = types.ModuleType("ai_trading.data.fetch")
 df_stub.get_bars = df_stub.get_bars_batch = lambda *a, **k: []
 df_stub.get_minute_df = lambda *a, **k: None
-sys.modules["ai_trading.data_fetcher"] = df_stub
+sys.modules["ai_trading.data.fetch"] = df_stub
 
 market_pkg = types.ModuleType("ai_trading.market")
 cal_stub = types.ModuleType("ai_trading.market.calendars")
