+++ tests/unit/test_data_fetcher_http.py
@@ -8,7 +8,7 @@ import pytest
 
 pd = pytest.importorskip("pandas")
 
-import ai_trading.data_fetcher as df
+import ai_trading.data.fetch as df
 
 
 class _Resp:
